requirements.txt
youtube-transcript-api
flask
pinecone-client
torch
torchvision
sentence-transformers
python-dotenv